+++ src/datasources/alarms/constants/AlarmsQueryEditor.constants.ts
@@ -1,13 +1,17 @@
 import { AlertVariant } from '@grafana/ui';
 import { AlarmsProperties } from '../types/ListAlarms.types';
+import { TransitionInclusionOption } from '../types/types';
 
 export const LABEL_WIDTH = 26;
 export const CONTROL_WIDTH = 65;
+export const SECONDARY_LABEL_WIDTH = 20;
+export const SECONDARY_CONTROL_WIDTH = 26;
 export const ERROR_SEVERITY_WARNING: AlertVariant = 'warning';
 export const QUERY_EDITOR_MIN_TAKE = 1;
 export const QUERY_EDITOR_MAX_TAKE = 10000;
 export const DEFAULT_QUERY_EDITOR_TAKE = 1000;
 export const DEFAULT_QUERY_EDITOR_DESCENDING = true;
+export const DEFAULT_QUERY_EDITOR_TRANSITION_INCLUSION_OPTION = TransitionInclusionOption.None;
 
 export const ALARMS_TIME_FIELDS = [
   AlarmsProperties.occurredAt,
@@ -23,6 +27,7 @@ export const labels = {
   properties: 'Properties',
   descending: 'Descending',
   take: 'Take',
+  transitionInclusion: 'Include Transition',
 };
 
 export const tooltips = {
@@ -30,7 +35,8 @@ export const tooltips = {
   queryBy: 'This optional field specifies the query filters.',
   properties: 'This field specifies the properties to use in the query.',
   descending: 'This toggle returns the alarms query in descending order.',
-  take: 'This field specifies the maximum number of alarms to return.'
+  take: 'This field specifies the maximum number of alarms to return.',
+  transitionInclusion: 'This field specifies whether to include all transitions, only the most recent, or none.',
 };
 
 export const takeErrorMessages = {
@@ -153,3 +159,24 @@ export const AlarmsPropertiesOptions: Record<
     value: AlarmsProperties.workspace,
   },
 }
+
+export const AlarmsTransitionInclusionOptions: Record<
+  TransitionInclusionOption,
+  {
+    label: string;
+    value: TransitionInclusionOption;
+  }
+> = {
+  [TransitionInclusionOption.None]: {
+    label: 'None',
+    value: TransitionInclusionOption.None,
+  },
+  [TransitionInclusionOption.MostRecentOnly]: {
+    label: 'Most recent only',
+    value: TransitionInclusionOption.MostRecentOnly,
+  },
+  [TransitionInclusionOption.All]: {
+    label: 'All',
+    value: TransitionInclusionOption.All,
+  },
+};